Adding 68 order to new branch
These are the original changes which have been removed from the master branch and added to a new branch.

Now able to read LCO spectra with the 68th oder. Switched to smart_shk functions temporarily to use full orders in flat files and spectra.

Breaks old spectras and requires fix to work on old spectra and new simultaniously.